requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
channels==2.1.7
channels-redis==2.3.3
constantly==15.1.0
daphne==2.2.5
Django==2.1.7
django-filter==2.1.0
django-model-utils==3.1.2
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
Markdown==3.0.1
more-itertools==6.0.0
msgpack==0.6.1
pluggy==0.9.0
py==1.8.0
PyHamcrest==1.9.0
pytest==4.3.1
pytest-asyncio==0.10.0
pytest-django==3.4.8
pytz==2018.9
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
ws4py==0.5.1
zope.interface==4.6.0
